Added top_comments.py and fixed bug
top_comments.py:

Moved the getItems method from the writing_prompts.py to this file
and then import this file into the writing_prompts.py to get the top
post of the day (be it a WP, PI or IP).
This seperation allows to get the top comment post from other subreddits
also.

fixed bug:

Clicking on 'Writing Prompts' in the landing page won't return an
error now. Previously I was automatically passing the subreddit name
from the route (.../writingprompts) as an arguement to the function call
when the get_writingprompt function that was being called did accept an
arguement in the first place.
This has now been fixed because I am  now importing and using
top_comments.py in writing_prompts.py .